requirements.txt
Django==1.11.5
djangorestframework==3.6.4
pytz==2017.2
psycopg2==2.7.3.1
raven==6.1.0
requests==2.18.4